Rename test classes to have a Facts suffix
Fixes #163
