+++ src/AdvApi32.Tests/AdvApi32.Tests.csproj
@@ -31,7 +31,7 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AdvApi32.cs" />
+    <Compile Include="AdvApi32Facts.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
