+++ src/AdvApi32.Tests/AdvApi32Facts.cs
@@ -2,11 +2,15 @@
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 
 using System;
-using System.IO;
 using PInvoke;
 using Xunit;
-using static PInvoke.Kernel32;
+using static PInvoke.AdvApi32;
 
-public partial class Kernel32
+public class AdvApi32Facts
 {
+    [Fact]
+    public void LsaNtStatusToWinError_UsesTable()
+    {
+        Assert.Equal(Win32ErrorCode.ERROR_NOACCESS, LsaNtStatusToWinError(NTSTATUS.Code.STATUS_ACCESS_VIOLATION));
+    }
 }
