+++ src/ImageHlp.Tests/ImageHlp.Tests.csproj
@@ -31,7 +31,7 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ImageHlp.cs" />
+    <Compile Include="ImageHlpFacts.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
