+++ src/Kernel32.Tests.Shared/Kernel32.Tests.Shared.projitems
@@ -10,7 +10,7 @@
     </Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Kernel32.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Kernel32Facts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NTStatusExceptionTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Win32ExceptionTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Kernel32ExtensionsTests.cs" />
