+++ src/Kernel32.Tests.Shared/Kernel32Facts.cs
@@ -2,14 +2,11 @@
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 
 using System;
+using System.IO;
 using PInvoke;
 using Xunit;
-using static PInvoke.SetupApi;
+using static PInvoke.Kernel32;
 
-public class SetupApi
+public partial class Kernel32Facts
 {
-    [Fact(Skip = "No tests yet")]
-    public void NoTests()
-    {
-    }
 }
