+++ src/MSCorEE.Tests/MSCorEE.Tests.csproj
@@ -33,7 +33,7 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MSCorEE.cs" />
+    <Compile Include="MSCorEEFacts.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
